move test client into client crate, removed unneeded tests from perf
