+++ rotala-client/tests/staticweight_test.rs
@@ -4,8 +4,9 @@ use rotala_client::broker::uist::UistBrokerBuilder;
 use rotala_client::broker::BrokerCost;
 
 use rotala::input::penelope::Penelope;
+use rotala_client::client::uist_v1::LocalClient;
 use rotala_client::strategy::staticweight::{PortfolioAllocation, StaticWeightStrategyBuilder};
-use rotala_http::http::uist_v1::{Client, TestClient};
+use rotala_http::http::uist_v1::Client;
 
 #[tokio::test]
 async fn staticweight_integration_test() {
@@ -19,7 +20,7 @@ async fn staticweight_integration_test() {
     weights.insert("BCD".to_string(), 0.5);
 
     let source = Penelope::random(length_in_days, vec!["ABC", "BCD"]);
-    let mut client = TestClient::single("Random", source);
+    let mut client = LocalClient::single("Random", source);
     let resp = client.init("Random".to_string()).await.unwrap();
 
     let brkr = UistBrokerBuilder::new()
